requirements.txt
requests
datetime
pandas
sqlalchemy
psycopg2-binary
python-dotenv
schedule